+++ src/components/MultiSeriesBarChart/utilities/format-aria-label.ts
@@ -1,0 +1,7 @@
+import {AccessibilitySeries} from '../types';
+
+export function formatAriaLabel(series: AccessibilitySeries) {
+  return `${series.title}: ${series.data.map(
+    ({label, value}) => `${label} ${value}`,
+  )}`;
+}