+++ src/components/MultiSeriesBarChart/utilities/index.ts
@@ -1,2 +1,3 @@
 export {getStackedValues} from './get-stacked-values';
 export {getMinMax} from './get-min-max';
+export {formatAriaLabel} from './format-aria-label';